split Homefragment into Quick View and Detail
