+++ app/src/main/res/layout/fragment_quick.xml
@@ -1,0 +1,7 @@
+<LinearLayout
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical"
+    android:id="@+id/quick_fragment">
+</LinearLayout>
